+++ hugin/src/protocol/mod.rs
@@ -1,8 +1,11 @@
 pub mod cmd;
 pub mod connection;
 pub mod frame;
+pub(crate) mod multisig;
 pub mod parse;
 pub mod shutdown;
+pub(crate) mod smartcontract;
+pub(crate) mod stdtx;
 
 pub use cmd::*;
 pub use connection::*;
