+++ hugin/src/protocol/smartcontract/mod.rs
@@ -1,0 +1,10 @@
+/*
+#################################################################################
+# See LICENSE.md for full license information.                                  #
+# Software: Drasil Blockchain Application Framework                             #
+# License: Drasil Source Available License v1.0                                 #
+# Licensors: Torben Poguntke (torben@drasil.io) & Zak Bassey (zak@drasil.io)    #
+#################################################################################
+*/
+pub(crate) mod nft_marketplace;
+pub(crate) use nft_marketplace::*;
