+++ hugin/src/protocol/smartcontract/nft_marketplace/mod.rs
@@ -1,0 +1,10 @@
+/*
+#################################################################################
+# See LICENSE.md for full license information.                                  #
+# Software: Drasil Blockchain Application Framework                             #
+# License: Drasil Source Available License v1.0                                 #
+# Licensors: Torben Poguntke (torben@drasil.io) & Zak Bassey (zak@drasil.io)    #
+#################################################################################
+*/
+//pub(crate) mod reward_handler;
+//pub(crate) use reward_handler::handle_rewardclaim;
