+++ murin/src/txbuilders/rwdist/mod.rs
@@ -12,7 +12,7 @@ pub use build_rwd::*;
 pub mod finalize_rwd;
 pub use finalize_rwd::*;
 
-pub mod build_utxopti;
+pub mod finalize_utxopti;
 
 use super::*;
 use cardano_serialization_lib as clib;
